Change buttons

+++ src/Pages/Admin/Forums/ViewForums.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import AdminNav from "../../../Components/AdminNav.component";
 import { Table } from 'reactstrap';
 import axios from 'axios';
-import firebase from "../../../Firebase";
+import SEButton from "./SEButton";
 var forums;
 
 export default class ViewForums extends Component {
@@ -14,8 +14,10 @@ export default class ViewForums extends Component {
             title: "",
             date: "",
             startDate: "",
-            startTime: "",  
+            startTime: "",
+            status: ""
         }
+        // this.startForum = this.startForum.bind(this);
     }
     
     //get data
@@ -27,7 +29,8 @@ export default class ViewForums extends Component {
                 title: forum.title,
                 date: forum.date,
                 startDate: forum.startDate,
-                startTime: forum.startTime
+                startTime: forum.startTime,
+                status: forum.status
             }));
             this.setState({
                 forums: res.data
@@ -35,62 +38,70 @@ export default class ViewForums extends Component {
         })
     }
 
-    //add document to firebase collection
-    // postEventID(id, e) {
+    // send start forum id to bakend
+    // startForum(id, e) {
+        
     //     e.preventDefault();
-    //     const db = firebase.firestore();
-    //     db.collection('messages').add({
-    //         eventId: id,
-    //         userRole: localStorage.getItem('erole')
+        
+    //     axios.post("http://localhost:8080/sendForumID/" +id)
+    //     .then((response) => {
+    //         console.log(response);
+    //     }).catch((error) => {
+    //         console.log(error);
     //     });
-    //     // this.setState({forums});
-    //     console.log("Added event " + id + " and user " + localStorage.getItem('erole') + " to firebase");
+    //     this.setState({
+    //         isStart: false
+    //     });
+    //     // this.btn.setAttribute("disabled", "disabled");
     // }
 
-    //send event id to bakend
-    postEventID(id, e) {
-        e.preventDefault();
-        this.setState({
-            loading: true
-        });
-        axios.post("http://localhost:8080/sendForumID", {
-            id: id
-        }).then((response) => {
-            console.log(response);
-        }).catch((error) => {
-            console.log(error);
-        });
-    }
 
-    //delete button
-    deleteForum(id, e) {
-       e.preventDefault();
-
-       axios.delete("http://localhost:8080/deleteForum/" +id)
-       .then(res => {
-            console.log(res);
-            console.log(res.data);
-            let forums = this.state.forums.filter(item => item.id !== id);
-            this.setState({forums});
-       });
-   }
+    // //send end forum id to backend
+    // endForums(id, e) {
+    //     e.preventDefault();
+        
+    //     axios.post("http://localhost:8080/endForumID/" +id)
+    //     .then((response) => {
+    //         console.log(response);
+    //     }).catch((error) => {
+    //         console.log(error);
+    //     });
+    //     // this.btn.setAttribute("disabled", "disabled");  
+    // }
 
    //first render
     renderForums() {
         return this.state.forums.map((forum, index) => {
-            const {id, title, date, startDate, startTime} = forum
+            const {id, title, date, startDate, startTime, status} = forum
             return (
                 <tr key={id}>
                     <td>{title}</td>
                     <td>{date}</td>
                     <td>{startDate}</td>
                     <td>{startTime}</td>
-                    <td><button onClick={(e) => this.postEventID(id, e)}>Start</button></td>
-                    <td><button onClick={null}>End</button></td>
-                    <td><button onClick={(e) => this.deleteForum(id, e)}>Delete</button></td>
+                    <SEButton color1={"green"} color2="{red}" name1={"Start"} name2={"End"} id={id}/>
+                    {/* <td><button ref={btn => {this.btn = btn; }} onClick={this.state.isStart ? (e) => this.startForum(id, e) : (e) => this.endForums(id, e)}>{this.state.isStart ? 'Start' : 'End'}</button></td> */}
+                    <td><button onClick={(e) => this.deleteForum(id, status, e)}>Delete</button></td>
                 </tr>
             )
         })
+    }
+
+    //delete button
+    deleteForum(id, status, e) {
+        e.preventDefault();
+        if(status == '0') {
+            axios.delete("http://localhost:8080/deleteForum/" +id)
+            .then(res => {
+                console.log(res);
+                console.log(res.data);
+                let forums = this.state.forums.filter(item => item.id !== id);
+                this.setState({forums});
+            });
+        }else {
+            console.log("Forum has already started");
+        }
+         
     }
 
     render(forums, index) {
@@ -108,7 +119,7 @@ export default class ViewForums extends Component {
                         </tr>
                     </thead>
                     <tbody>
-                       {this.renderForums()}
+                        {this.renderForums()}
 
                     </tbody>
                 </Table>
@@ -116,4 +127,6 @@ export default class ViewForums extends Component {
         );
     }
 }
+
+
